cleaned up and added magic-ville link importer
